refactor(rust): share the wallet JWT builder between cdp and openfort
Openfort's inline copy of the CDP wallet-JWT code (claims struct,
jwt_uri, extract_host, sort_json, compute_req_hash, ES256 encode) moves
to a shared wallet_jwt module parameterized by provider name. Each
backend keeps only its own secret-format handling: CDP the base64
PKCS#8 DER decode, Openfort the PEM-or-bare-base64 normalization.
Duplicated jwt_uri/req-hash tests collapse into the shared module's
test file.
