+++ rust/src/cdp/jwt.rs
@@ -1,10 +1,10 @@
 use crate::error::SignerError;
 use crate::sdk_adapter::{keypair_from_seed, keypair_pubkey};
+use crate::wallet_jwt;
 use base64::{engine::general_purpose::STANDARD, Engine};
 use jsonwebtoken::{encode, Algorithm, EncodingKey, Header};
 use serde::Serialize;
 use serde_json::Value;
-use sha2::{Digest, Sha256};
 use uuid::Uuid;
 
 /// PKCS#8 DER prefix for Ed25519 private keys.
@@ -23,17 +23,6 @@ struct AuthClaims {
     nbf: i64,
     exp: i64,
     uris: Vec<String>,
-}
-
-#[derive(Serialize)]
-struct WalletClaims {
-    uris: Vec<String>,
-    iat: i64,
-    nbf: i64,
-    exp: i64,
-    jti: String,
-    #[serde(rename = "reqHash", skip_serializing_if = "Option::is_none")]
-    req_hash: Option<String>,
 }
 
 /// Wraps raw DER bytes as a PKCS#8 PEM string (-----BEGIN PRIVATE KEY-----).
@@ -76,69 +65,6 @@ fn validate_ed25519_keypair_bytes(key_bytes: &[u8]) -> Result<(), SignerError> {
     Ok(())
 }
 
-/// Build the URI claim value for a JWT.
-pub(super) fn jwt_uri(host: &str, method: &str, path: &str) -> String {
-    format!("{method} {host}{path}")
-}
-
-/// Extract request host (including port if present) from a base URL.
-pub(super) fn extract_host(base_url: &str) -> Result<String, SignerError> {
-    let url = reqwest::Url::parse(base_url)
-        .map_err(|_| SignerError::ConfigError(format!("Invalid CDP base URL: {base_url}")))?;
-
-    let host = url.host_str().ok_or_else(|| {
-        SignerError::ConfigError(format!("Missing host in CDP base URL: {base_url}"))
-    })?;
-
-    Ok(match url.port() {
-        Some(port) => format!("{host}:{port}"),
-        None => host.to_string(),
-    })
-}
-
-/// Recursively sort JSON object keys for deterministic hashing.
-fn sort_json(value: &Value) -> Value {
-    match value {
-        Value::Object(map) => {
-            let mut keys: Vec<&String> = map.keys().collect();
-            keys.sort();
-            let mut sorted = serde_json::Map::with_capacity(map.len());
-            for key in keys {
-                if let Some(value) = map.get(key) {
-                    sorted.insert(key.clone(), sort_json(value));
-                }
-            }
-            Value::Object(sorted)
-        }
-        Value::Array(values) => Value::Array(values.iter().map(sort_json).collect()),
-        _ => value.clone(),
-    }
-}
-
-/// Compute the request body hash for wallet authentication, if required.
-pub(super) fn compute_req_hash(body: Option<&Value>) -> Result<Option<String>, SignerError> {
-    let body = match body {
-        Some(body) => body,
-        None => return Ok(None),
-    };
-
-    if body.is_null() {
-        return Ok(None);
-    }
-
-    if matches!(body, Value::Object(map) if map.is_empty()) {
-        return Ok(None);
-    }
-
-    let sorted = sort_json(body);
-    let json = serde_json::to_string(&sorted).map_err(|e| {
-        SignerError::SerializationError(format!("Failed to serialize request body: {e}"))
-    })?;
-
-    let hash = Sha256::digest(json.as_bytes());
-    Ok(Some(hex::encode(hash)))
-}
-
 /// Generate a random nonce for JWT header fields.
 fn random_nonce() -> String {
     Uuid::new_v4().simple().to_string()
@@ -164,7 +90,7 @@ pub(super) fn create_auth_jwt(
         iat: now,
         nbf: now,
         exp: now + 120,
-        uris: vec![jwt_uri(host, method, path)],
+        uris: vec![wallet_jwt::jwt_uri(host, method, path)],
     };
 
     if api_key_secret.starts_with("-----BEGIN") {
@@ -220,19 +146,6 @@ pub(super) fn create_wallet_jwt(
     path: &str,
     request_body: Option<&Value>,
 ) -> Result<String, SignerError> {
-    let now = chrono::Utc::now().timestamp();
-
-    let req_hash = compute_req_hash(request_body)?;
-
-    let claims = WalletClaims {
-        uris: vec![jwt_uri(host, method, path)],
-        iat: now,
-        nbf: now,
-        exp: now + 120,
-        jti: Uuid::new_v4().to_string(),
-        req_hash,
-    };
-
     // The wallet secret is base64-encoded PKCS#8 DER (P-256)
     let der_bytes = STANDARD.decode(wallet_secret).map_err(|_e| {
         #[cfg(feature = "unsafe-debug")]
@@ -248,12 +161,5 @@ pub(super) fn create_wallet_jwt(
         SignerError::InvalidPrivateKey("Failed to parse walletSecret as EC private key".to_string())
     })?;
 
-    let mut header = Header::new(Algorithm::ES256);
-    header.typ = Some("JWT".to_string());
-
-    encode(&header, &claims, &key).map_err(|_e| {
-        #[cfg(feature = "unsafe-debug")]
-        log::error!("Failed to encode wallet JWT: {_e}");
-        SignerError::SigningFailed("Failed to create CDP wallet JWT".to_string())
-    })
+    wallet_jwt::create_wallet_jwt("CDP", &key, host, method, path, request_body)
 }
